Extract tracking number information into arrays
Instead of giving back a string for tracking numbers we
return arrays instead.  This is a backwards-incompatible change,
and as such requires a major version update.  All current fulfillment
providers, and future providers will need to providing tracking
numbers as arrays, even if it is guaranteed that they only provide
a single tracking number.